requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
crewai>=0.28.0
openai>=1.3.0